Migration checks (#597)
* Initial commit of the turbot migration checks

* Skip the `turbot_lockdown*` policies.
They are supposed to be attached to users and roles.  We don't care if they are there.

* Skip the `turbot_lockdown*` policies.
They are supposed to be attached to users and roles.  We don't care if they are there.

* Skip the `turbot_lockdown*` policies.
They are supposed to be attached to users and roles.  We don't care if they are there.

* Skip the `turbot_lockdown*` policies.
They are supposed to be attached to users and roles.  We don't care if they are there.

* Add account_id to the iam_users and iam_groups dump
We need the account id for all these users and groups to build one big permissions verification file instead of separate files for each account.